Move more functionality from QuCircuit to CircuitStructure and CircuitSetter
